helpmecorona/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.11
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==19.9.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.3
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
